refactor: update imports to use shared components for consistency and maintainability
