+++ src/app/(toys)/congrats-button/page.mdx
@@ -1,5 +1,5 @@
 import { CongratsButton } from "./_components/congrats-button";
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 # おめでとうボタン
