+++ src/app/(toys)/device-orientation/page.mdx
@@ -1,5 +1,5 @@
 import { DeviceOrientationSample } from "./_components/device-orientation-sample";
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 # 傾き検知サンプル集
