+++ src/app/(toys)/euclidean-algorithm/page.mdx
@@ -1,5 +1,5 @@
 import { GcdCalc } from "./_components/gcd-calc";
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 # ユークリッドの互除法
