+++ src/app/(toys)/fizz-buzz/page.mdx
@@ -1,5 +1,5 @@
 import { FizzBuzzCalc } from "./_components/fizz-buzz-calc";
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 # Fizz Buzz
