+++ src/app/(toys)/hello-world/markdown/page.mdx
@@ -166,12 +166,12 @@ The source file type is ".md**x**". So you can use JavaScript and React.Componen
 ### Call Components
 
 ```mdx
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 <Version version="0.1.0" onDate="1970-01-01" />
 ```
 
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 <Version version="0.1.0" onDate="1970-01-01" />
