+++ src/app/(toys)/modane-3d-viewer/page.mdx
@@ -1,5 +1,5 @@
 import { Modane3dViewer } from "./_components/modane-3d-viewer";
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 # もだねちゃん 3D ビューワー
