+++ src/app/(toys)/nabeatsu-assessmenter/page.mdx
@@ -1,5 +1,5 @@
 import { NabeatsuAssessmenter } from "./_components/nabeatsu-assessmenter";
-import { Version } from "@/components/atoms/version";
+import { Version } from "@/components/shared/version";
 
 # ナベアツ判定機
