+++ src/components/shared/labeled-number-input-old.tsx
