+++ src/components/shared/result-display.tsx
